Module Current_incr__Modifiable.Separate

Parameters

Map : Stdlib.Map.S

Signature

val map : unit Map.t t -> (Map.key -> 'b t -> changeable) -> 'b Map.t t

map x fn applies fn to each element of x and returns a map from input elements to results. When new elements are added to x, it only runs fn on the new elements, rather than on all elements.