Module Current_incr__Time

Basic ordering operations

type t

An element of an ordering.

val root : unit -> t

Create a new ordering with a single element. O(1)

val after : ?⁠on_forget:(unit -> unit) -> t -> t

after t inserts a new element to the ordering, greater than t but less than all existing elements greater than t.

O(1) amortized.

val same_order : t -> t -> bool

Check if two elements belong to the same order. O(1)

val compare : t -> t -> int

Compare two elements. O(1)

val cardinal : t -> int

How many elements are ordered. O(1)

Memory management

val forget : t -> unit

Memory of every element is retained. When you know you are not going to use an element any longer, forget it to release memory. O(1).

val is_valid : t -> bool

After calling forget, an element should not be used. You can check if it is the case with is_valid.

val splice_out : t -> t -> unit

splice_out ts te forgets all times between ts and te, so that afterwards te immediately follows ts.

val set_forget : t -> (unit -> unit) -> unit

set_forget t fn sets t's forget function to fn. This is called when t is forgotten. It is an error if t already has a forget function.

val clear_forget : t -> unit

clear_forget t reverses the effect of set_forget.

val next : t -> t

next t is the time immediately after t (which must not be the last time).

val prev : t -> t

prev t is the time immediately before t (which must not be the first time).